% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cross_validation_logistic}
\alias{cross_validation_logistic}
\title{Cross-validation for logistic regression}
\usage{
cross_validation_logistic(X, y, K, M)
}
\arguments{
\item{X}{a p x n matrix of regressor}

\item{y}{a n x 1 vector of response}

\item{K}{number of splits}

\item{M}{number of repetitions}
}
\description{
Cross-validation for logistic regression
}
