% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cross_validation_logistic_count}
\alias{cross_validation_logistic_count}
\title{Cross-validation for logistic regression with counting error}
\usage{
cross_validation_logistic_count(X, y, seed, K = 10L, M = 10L)
}
\arguments{
\item{X}{a n x p matrix of regressor}

\item{y}{a n-vector of response}

\item{seed}{an integer for setting the seed (reproducibility)}

\item{K}{number of splits; 10 by default}

\item{M}{number of repetitions; 10 by default}
}
\description{
Cross-validation for logistic regression with counting error
}
