% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_reg}
\alias{log_reg}
\title{Logistic regression}
\usage{
log_reg(X, y)
}
\arguments{
\item{X}{a p x n matrix of regressor}

\item{y}{a n x 1 vector of response}
}
\description{
Logistic regression
}
