% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logistic_mle}
\alias{logistic_mle}
\title{Logistic regression based on MLE}
\usage{
logistic_mle(X, y)
}
\arguments{
\item{X}{a n x p matrix of regressor}

\item{y}{a n-vector of response}
}
\description{
Logistic regression based on MLE
}
