% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pred_log_reg}
\alias{pred_log_reg}
\title{Prediction of logistic regression}
\usage{
pred_log_reg(X, y, X_new)
}
\arguments{
\item{X}{a p x n matrix of regressor}

\item{y}{a n x 1 vector of response}

\item{X_new}{a p x m matrix of new regressor}
}
\description{
Prediction of logistic regression
}
